+++ axum/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 # Unreleased
 
 - **fixed:** Improve `debug_handler` on tuple response types ([#2201])
+- **breaking:** Make `serve` generic over the listener and IO types
 
 [#2201]: https://github.com/tokio-rs/axum/pull/2201